SPUI-1453 > Disclaimer (#638)
* SPUI-1455 > add disclaimer

* SPUI-1455 > add disclaimer

* SPUI-1455 > disclaimer

* SPUI-1455 > add disclaimers to router

* SPUI-1455 > add disclaimers to router

* SPUI-1455 > add disclaimers to router

* SPUI-1455 > Add logic to continue button on disclaimers popup

* SPUI-1455 > Add logic to continue button on disclaimers popup

* SPUI-1455 > fix ui

* SPUI-1455 > add single scroll to popup

* SPUI-1455 > add parser to link

* SPUI-1455 > fix model

* SPUI-1455 > fix questions ui

* SPUI-1455 > fix questions ui

* SPUI-1455 > fix questions ui

* SPUI-1455 > fix questions ui

* SPUI-1455 > fix comments

* SPUI-1455 > fix scrolling height box

+++ lib/service/services/signal_r/model/recurring_buys_model.dart
@@ -11,8 +11,8 @@ class RecurringBuysModel with _$RecurringBuysModel {
   const factory RecurringBuysModel({
     double? fromAmount,
     String? scheduledTime,
-    int? scheduledDayOfWeek,
-    int? scheduledDayOfMonth,
+    double? scheduledDayOfWeek,
+    double? scheduledDayOfMonth,
     String? nextExecution,
     String? lastExecutionTime,
     String? id,
